Move hop tests to separate module
